requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-rest-multiple-models==2.1.3
djangorestframework==3.12.4
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
waitress==2.0.0
